Create separate files for the classes in CoumpoundCRS file.
