+++ include/CRSBase.h
@@ -1,0 +1,32 @@
+#pragma once
+
+#include <spine/ConfigBase.h>
+
+namespace SmartMet
+{
+namespace Plugin
+{
+namespace WCS
+{
+class CRSBase
+{
+ public:
+  using Identifier = std::string;
+  using Abbrev = std::string;
+  CRSBase() {}
+  virtual ~CRSBase() {}
+  inline const Abbrev& getAbbrev() const { return mAbbrev; }
+  inline const Identifier& getIdentifier() const { return mIdentifier; }
+
+ protected:
+  void parse(boost::shared_ptr<SmartMet::Spine::ConfigBase> config, libconfig::Setting& setting);
+  void setAbbrev(const Abbrev& abbrev);
+  void setIdentifier(const Identifier& identifier);
+
+ private:
+  Identifier mIdentifier;
+  Abbrev mAbbrev;
+};
+}
+}
+}
